requirements.txt
langchain_core
langchain_community
langchain_openai
python-dotenv
streamlit



